+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/dateam/toolkit/compare/v0.1.0...v0.1.3) (2020-12-01)
+
+**Note:** Version bump only for package @dateam/ark
+
+
+
+
+
+## [0.1.2](https://github.com/dateam/toolkit/compare/v0.1.0...v0.1.2) (2020-12-01)
+
+**Note:** Version bump only for package @dateam/ark
+
+
+
+
+
 # 0.1.0 (2020-11-05)
 
 **Note:** Version bump only for package @dateam/ark